restful: decorator improvement
* Improves require_api_auth decorator by clearing the session
  and returning the response at the end.

* Removes the bound_instance_ parameter, removes the
  require.oauth from the return statement and the returned
  function now contains only the *args and the **kwargs.

+++ invenio/modules/tags/restful.py
@@ -139,17 +139,21 @@ class TagResource(Resource):
         error_handler
     ]
 
-    def get(self, oauth, tag_name):
+    def get(self, tag_name):
         """Get a tag.
 
         :param tag_name: the name of the tag to retrieve
         """
-        uid = current_user.get_id()
+        if 'access_token=' not in request.url:
+            raise TagOwnerError(error_msg="Unauthorized", status_code=401)
+        access_token = request.url[request.url.find('=')+1:len(request.url)]
+        uid = tags_api.get_user_id_from_access_token(access_token)
+        # uid = current_user.get_id()
         tag_retrieved = tags_api.get_tag_of_user(uid, tag_name)
         tag = TagRepresenation(tag_retrieved)
         return tag.marshal()
 
-    def delete(self, oauth, tag_name):
+    def delete(self, tag_name):
         """Delete a tag.
 
         Checks if the tag is attached to records. If True,
@@ -157,12 +161,16 @@ class TagResource(Resource):
 
         :param tag_name: the name of the tag to delete
         """
-        uid = current_user.get_id()
+        if 'access_token=' not in request.url:
+            raise TagOwnerError(error_msg="Unauthorized", status_code=401)
+        access_token = request.url[request.url.find('=')+1:len(request.url)]
+        uid = tags_api.get_user_id_from_access_token(access_token)
+        # uid = current_user.get_id()
         tags_api.delete_tag_from_user(uid, tag_name)
         return "", 204
 
     @require_header('Content-Type', 'application/json')
-    def patch(self, oauth, tag_name):
+    def patch(self, tag_name):
         """Update a tag.
 
         The attributes that can be updated are:
@@ -178,24 +186,28 @@ class TagResource(Resource):
                 error_msg="Validation for tag update failed",
                 status_code=400,
                 error_list=v.get_errors())
-        uid = current_user.get_id()
+        if 'access_token=' not in request.url:
+            raise TagOwnerError(error_msg="Unauthorized", status_code=401)
+        access_token = request.url[request.url.find('=')+1:len(request.url)]
+        uid = tags_api.get_user_id_from_access_token(access_token)
+        # uid = current_user.get_id()
         tag_retrieved = tags_api.update_tag_of_user(uid, tag_name, json_data)
         tag = TagRepresenation(tag_retrieved)
         return tag.marshal(), 201
 
-    def post(self, oauth, tag_name):
+    def post(self, tag_name):
         """post."""
         abort(405)
 
-    def options(self, oauth, tag_name):
+    def options(self, tag_name):
         """options."""
         abort(405)
 
-    def put(self, oauth, tag_name):
+    def put(self, tag_name):
         """put."""
         abort(405)
 
-    def head(self, oauth, tag_name):
+    def head(self, tag_name):
         """head."""
         abort(405)
 
@@ -209,27 +221,35 @@ class TagListResource(Resource):
         error_handler
     ]
 
-    def get(self, oauth):
+    def get(self):
         """ Get a list of tags.
 
         Get the list of tags a user owns.
         """
-        uid = current_user.get_id()
+        if 'access_token=' not in request.url:
+            raise TagOwnerError(error_msg="Unauthorized", status_code=401)
+        access_token = request.url[request.url.find('=')+1:len(request.url)]
+        uid = tags_api.get_user_id_from_access_token(access_token)
+        # uid = current_user.get_id()
         tags_retrieved = tags_api.get_all_tags_of_user(uid)
         tags = [TagRepresenation(t) for t in tags_retrieved]
         return map(lambda t: t.marshal(), tags)
 
-    def delete(self, oauth):
+    def delete(self):
         """Delete all tags.
 
         Delete all the tags a user owns.
         """
-        uid = current_user.get_id()
+        if 'access_token=' not in request.url:
+            raise TagOwnerError(error_msg="Unauthorized", status_code=401)
+        access_token = request.url[request.url.find('=')+1:len(request.url)]
+        uid = tags_api.get_user_id_from_access_token(access_token)
+        # uid = current_user.get_id()
         tags_api.delete_all_tags_from_user(uid)
         return "", 204
 
     @require_header('Content-Type', 'application/json')
-    def post(self, oauth):
+    def post(self):
         """Create a new tag.
 
         Creates a new tag and sets as owner the current user.
@@ -241,24 +261,28 @@ class TagListResource(Resource):
                 error_msg="Validation error for tag creation",
                 status_code=400,
                 error_list=v.get_errors())
-        uid = current_user.get_id()
+        if 'access_token=' not in request.url:
+            raise TagOwnerError(error_msg="Unauthorized", status_code=401)
+        access_token = request.url[request.url.find('=')+1:len(request.url)]
+        uid = tags_api.get_user_id_from_access_token(access_token)
+        # uid = current_user.get_id()
         tag_to_create = tags_api.create_tag_for_user(uid, json_data['name'])
         tag_to_return = TagRepresenation(tag_to_create)
         return tag_to_return.marshal(), 201
 
-    def patch(self, oauth):
+    def patch(self):
         """PATCH."""
         abort(405)
 
-    def options(self, oauth):
+    def options(self):
         """OPTIONS."""
         abort(405)
 
-    def put(self, oauth):
+    def put(self):
         """PUT."""
         abort(405)
 
-    def head(self, oauth):
+    def head(self):
         """HEAD."""
         abort(405)
 
@@ -277,17 +301,21 @@ class RecordTagResource(Resource):
         error_handler
     ]
 
-    def delete(self, oauth, record_id, tag_name):
+    def delete(self, record_id, tag_name):
         """Detach a tag from a record.
 
         :param record_id: the identifier of the record
         :param tag_name: the name of the tag
         """
-        uid = current_user.get_id()
+        if 'access_token=' not in request.url:
+            raise TagOwnerError(error_msg="Unauthorized", status_code=401)
+        access_token = request.url[request.url.find('=')+1:len(request.url)]
+        uid = tags_api.get_user_id_from_access_token(access_token)
+        # uid = current_user.get_id()
         tags_api.detach_tag_from_record(uid, tag_name, record_id)
         return "", 204
 
-    def post(self, oauth, record_id, tag_name):
+    def post(self, record_id, tag_name):
         """A POST request.
 
         :param record_id: the identifier of the record
@@ -295,7 +323,7 @@ class RecordTagResource(Resource):
         """
         abort(405)
 
-    def put(self, oauth, record_id, tag_name):
+    def put(self, record_id, tag_name):
         """A PUT request.
 
         :param record_id: the identifier of the record
@@ -303,7 +331,7 @@ class RecordTagResource(Resource):
         """
         abort(405)
 
-    def patch(self, oauth, record_id, tag_name):
+    def patch(self, record_id, tag_name):
         """A PATCH request.
 
         :param record_id: the identifier of the record
@@ -311,7 +339,7 @@ class RecordTagResource(Resource):
         """
         abort(405)
 
-    def options(self, oauth, record_id, tag_name):
+    def options(self, record_id, tag_name):
         """A OPTIONS request.
 
         :param record_id: the identifier of the record
@@ -319,7 +347,7 @@ class RecordTagResource(Resource):
         """
         abort(405)
 
-    def head(self, oauth, record_id, tag_name):
+    def head(self, record_id, tag_name):
         """A HEAD request.
 
         :param record_id: the identifier of the record
@@ -327,7 +355,7 @@ class RecordTagResource(Resource):
         """
         abort(405)
 
-    def get(self, oauth, record_id, tag_name):
+    def get(self, record_id, tag_name):
         """A GET request.
 
         :param record_id: the identifier of the record
@@ -346,7 +374,7 @@ class RecordListTagResource(Resource):
     ]
 
     @require_header('Content-Type', 'application/json')
-    def post(self, oauth, record_id):
+    def post(self, record_id):
         """Attach a list of tags to a record.
 
         If a tag in the list exists in database then it is attached
@@ -361,7 +389,9 @@ class RecordListTagResource(Resource):
                 error_msg="Validation error in attaching tags on record",
                 status_code=400,
                 error_list=attachTagsValidator.get_errors())
-        tags_just_attached = tags_api.attach_tags_to_record(json_data['tags'],
+        uid = current_user.get_id()
+        tags_just_attached = tags_api.attach_tags_to_record(uid,
+                                                            json_data['tags'],
                                                             record_id)
         if len(tags_just_attached) == 0:
             return []
@@ -370,7 +400,7 @@ class RecordListTagResource(Resource):
                 lambda t: TagRepresenation(t).marshal(), tags_just_attached
             )
 
-    def get(self, oauth, record_id):
+    def get(self, record_id):
         """Retrieve all the attached on a record tags.
 
         :param record_id: the identifier of the record
@@ -381,29 +411,29 @@ class RecordListTagResource(Resource):
         else:
             return map(lambda t: TagRepresenation(t).marshal(), attached_tags)
 
-    def delete(self, oauth, record_id):
+    def delete(self, record_id):
         """Detach all the tags from a record.
 
         :param record_id: the identifier of the record
         """
         pass
 
-    def put(self, oauth, record_id):
+    def put(self, record_id):
         """Replace all tags for a record.
 
         :param record_id: the identifier of the record
         """
         pass
 
-    def head(self, oauth, record_id):
+    def head(self, record_id):
         """A HEAD request."""
         abort(405)
 
-    def patch(self, oauth, record_id):
+    def patch(self, record_id):
         """A PATCH request."""
         abort(405)
 
-    def options(self, oauth, record_id):
+    def options(self, record_id):
         """A OPTIONS request."""
         abort(405)
